* Add portable getprotobyname internal API. * Fix HTTP & SNMP modules for pgm_peer_t changes. * Fix undefined tpdu_length in send_odatav. * Update non-C89 net send to be Windows aware. * Fix building pgmdump with gettext. * Add restrict to HTTP module. * Allow NETDB to be specified with PGM_NETDB environment variable.

+++ openpgm/pgm/checksum_unittest.c
@@ -80,8 +80,8 @@ START_TEST (test_inet_pass_001)
 	guint16 csum = pgm_inet_checksum (source, sizeof(source), 0);
 /* function calculates answer in host order */
 	csum = g_htons (csum);
-	g_message ("IP checksum of \"%s\" (%d) is %u (%u)",
-		source, sizeof(source), csum, answer);
+	g_message ("IP checksum of \"%s\" (%u) is %u (%u)",
+		source, (unsigned)sizeof(source), csum, answer);
 
 	fail_unless (answer == csum, "checksum mismatch");
 }
